requirements.txt
asgiref==3.2.10
certifi==2019.9.11
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
MarkupSafe==1.1.1
more-itertools==8.7.0
packaging==20.9
pipenv==2018.11.26
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
Werkzeug==1.0.1
whitenoise==5.0.1
zc.lockfile==2.0

Faker~=8.1.0